Add files via upload
Added functionality for the objects to use the hashing function and sort themselves into the hashing table, then verified that the table worked as expected. Created a few functions for handling the hash table safely, to avoid the system crashing. Next step is finding a better hash function and creating a basic front-end to continuously accept inputs.